RFID Reader Detector Buzzer Code
The analog input, which reads the electromagnetic waves of RFID readers and converts to voltage, triggers a LilyPad Buzzer. When the converted voltage is greater than 1 the Buzzer plays a tone, signalling the presence or attempt of an RFID reader.

RFID Reader Detector electronics project and code created for DESN37900 at Sheridan College © 2018